HV: refine is_vm0() and add get_sos_vm()
Under sharing mode, VM0 is identical with SOS VM. But the coupling of
SOS VM and VM 0 is not friendly for partition mode. Idealy, we should
get SOS VM by its type. This patch creates a new get_sos_vm() api to
leave space for future use;

We still need a separate patch to rename the term of is_vm0 to is_sos_vm
in the future;

Signed-off-by: Victor Sun <victor.sun@intel.com>